+++ src/extension.ts
@@ -21,6 +21,25 @@ enum UUIDVersion {
 type UUIDCase = "lower" | "upper";
 type UUIDMultiCursorBehavior = "unique" | "repeat";
 
+const namespaceIDs = [
+  {
+    label: "DNS",
+    description: uuid.v5.DNS,
+  },
+  {
+    label: "URL",
+    description: uuid.v5.URL,
+  },
+  {
+    label: "ISO OID",
+    description: "6ba7b812-9dad-11d1-80b4-00c04fd430c8",
+  },
+  {
+    label: "X.500 DN",
+    description: "6ba7b814-9dad-11d1-80b4-00c04fd430c8",
+  }
+]
+
 function getUUIDv35Name() {
   return window.showInputBox({
     prompt: l10n.t("Enter a name value for the UUID"),
@@ -33,25 +52,13 @@ function getUUIDv35Name() {
   });
 }
 
-async function getUUIDv35Namespace() {
+async function getUUIDv35Namespace(uuidCase: UUIDCase) {
   const ret = await window.showQuickPick(
     [
-      {
-        label: "DNS",
-        description: uuid.v5.DNS,
-      },
-      {
-        label: "URL",
-        description: uuid.v5.URL,
-      },
-      {
-        label: "ISO OID",
-        description: "6ba7b812-9dad-11d1-80b4-00c04fd430c8",
-      },
-      {
-        label: "X.500 DN",
-        description: "6ba7b814-9dad-11d1-80b4-00c04fd430c8",
-      },
+      ...namespaceIDs.map(el => ({
+        label: el.label,
+        description: (uuidCase === "upper") ? el.description.toUpperCase() : el.description
+      })),
       {
         label: l10n.t("Custom ..."),
         description: l10n.t("Enter your own UUID namespace"),
@@ -98,18 +105,22 @@ function getUUIDVersion() {
   );
 }
 
-async function generateUUID(version: UUIDVersion) {
+async function generateUUID(version: UUIDVersion, uuidCase: UUIDCase) {
+  let ret: string;
+
   if (version === "v3" || version === "v5") {
-    const namespace = await getUUIDv35Namespace();
+    const namespace = await getUUIDv35Namespace(uuidCase);
     if (namespace == null) return;
 
     const name = await getUUIDv35Name();
     if (name == null) return;
 
-    return uuid[version](name, namespace);
+    ret = uuid[version](name, namespace);
+  } else {
+    ret = (uuid[version] as () => string)();
   }
 
-  return (uuid[version] as () => string)();
+  return (uuidCase === "upper") ? ret.toUpperCase() : ret;
 }
 
 async function innerHandler(
@@ -118,15 +129,13 @@ async function innerHandler(
   uuidCase: UUIDCase,
   multiCursorBehavior: UUIDMultiCursorBehavior
 ) {
-  let newUUID = await generateUUID(uuidVersion);
+  let newUUID = await generateUUID(uuidVersion, uuidCase);
   for (const [index, selection] of textEditor.selections.entries()) {
     if (index > 0 && multiCursorBehavior == "unique") {
-      newUUID = await generateUUID(uuidVersion);
+      newUUID = await generateUUID(uuidVersion, uuidCase);
     }
 
     if (newUUID == null) return;
-
-    if (uuidCase === "upper") newUUID = newUUID.toUpperCase();
 
     await textEditor.edit((edit) => {
       selection.isEmpty
@@ -140,8 +149,8 @@ async function innerHandler(
 
 // we can't use the provided TextEditorEdit as the new value is computed asynchronously.
 async function generateUUIDHandler(textEditor: TextEditor) {
-  const configuration = workspace.getConfiguration("vscodeUUID");
-  const uuidVersion = configuration.get("UUIDVersion") as UUIDVersion;
+  const configuration = workspace.getConfiguration("vscode-uuid");
+  const uuidVersion = configuration.get("defaultVersion") as UUIDVersion;
   const uuidCase = configuration.get("case") as UUIDCase;
   const multiCursorBehavior = configuration.get(
     "multiCursorBehavior"
@@ -152,7 +161,7 @@ async function generateUUIDHandler(textEditor: TextEditor) {
 
 // we can't use the provided TextEditorEdit as the new value is computed asynchronously.
 async function generateUUIDWithVersionHandler(textEditor: TextEditor) {
-  const configuration = workspace.getConfiguration("vscodeUUID");
+  const configuration = workspace.getConfiguration("vscode-uuid");
   const uuidCase = configuration.get("case") as UUIDCase;
   const multiCursorBehavior = configuration.get(
     "multiCursorBehavior"
@@ -164,29 +173,84 @@ async function generateUUIDWithVersionHandler(textEditor: TextEditor) {
   await innerHandler(textEditor, uuidVersion, uuidCase, multiCursorBehavior);
 }
 
-function generateNilUUID(textEditor: TextEditor, edit: TextEditorEdit) {
-  for (const selection of textEditor.selections) {
-    if (selection.isEmpty) {
-      edit.insert(selection.active, uuid.NIL);
-    } else {
-      edit.replace(selection, uuid.NIL);
+function insertUUID(value: string) {
+  const configuration = workspace.getConfiguration("vscode-uuid");
+  const uuidCase = configuration.get("case") as UUIDCase;
+
+  const toInsert = (uuidCase === "upper") ? value.toUpperCase() : value;
+
+  return (textEditor: TextEditor, edit: TextEditorEdit) => {
+    for (const selection of textEditor.selections) {
+      if (selection.isEmpty) {
+        edit.insert(selection.active, toInsert);
+      } else {
+        edit.replace(selection, toInsert);
+      }
     }
+  }
+}
+
+async function insertWellKnownUUID(textEditor: TextEditor) {
+  const configuration = workspace.getConfiguration("vscode-uuid");
+  const uuidCase = configuration.get("case") as UUIDCase;
+
+  const value = await window.showQuickPick(
+    [
+      ...namespaceIDs.map(el => ({
+        label: l10n.t('{0} namespace ID', el.label),
+        description: (uuidCase === "upper") ? el.description.toUpperCase() : el.description
+      })),
+      {
+        label: l10n.t("Nil"),
+        description: uuid.NIL,
+      },
+      {
+        label: l10n.t("Max"),
+        description: (uuidCase === "upper") ? uuid.MAX.toUpperCase() : uuid.MAX,
+      },
+    ],
+    {
+      placeHolder: l10n.t("Well-known UUID"),
+      canPickMany: false,
+    }
+  );
+
+  if (value == null) return;
+
+  const toInsert = (uuidCase === "upper") ? value.description.toUpperCase() : value.description;
+
+  for (const selection of textEditor.selections) {
+    await textEditor.edit((edit) => {
+      selection.isEmpty
+        ? // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+          edit.insert(selection.active, toInsert!)
+        : // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+          edit.replace(selection, toInsert!);
+    });
   }
 }
 
 export function activate(context: ExtensionContext) {
   context.subscriptions.push(
     commands.registerTextEditorCommand(
-      "vscodeUUID.generateUUID",
+      "vscode-uuid.generateUUID",
       generateUUIDHandler
     ),
     commands.registerTextEditorCommand(
-      "vscodeUUID.generateUUIDWithVersion",
+      "vscode-uuid.generateUUIDWithVersion",
       generateUUIDWithVersionHandler
     ),
     commands.registerTextEditorCommand(
-      "vscodeUUID.generateNilUUID",
-      generateNilUUID
+      "vscode-uuid.insertNilUUID",
+      insertUUID(uuid.NIL)
+    ),
+    commands.registerTextEditorCommand(
+      "vscode-uuid.insertMaxUUID",
+      insertUUID(uuid.MAX)
+    ),
+    commands.registerTextEditorCommand(
+      "vscode-uuid.insertWellKnownUUID",
+      insertWellKnownUUID
     )
   );
 }